level-2-opt-assistant/requirements.txt
groq
python-dotenv
rich
pydantic